WIP: initial split and deletion of parsers.py
